Rename jit_function to compiled-function.
