+++ sourcepawn/jit/AMBuilder
@@ -30,9 +30,9 @@ if binary.compiler.cc.behavior == 'msvc':
 
 binary.sources += [
   'BaseRuntime.cpp',
-  'engine2.cpp',
+  'compiled-function.cpp',
   'dll_exports.cpp',
-  'jit_function.cpp',
+  'engine2.cpp',
   'sp_vm_basecontext.cpp',
   'sp_vm_engine.cpp',
   'sp_vm_function.cpp',
